refactor: renamed namespaces to match the AdSecCore
