+++ AdSecCore/Builders/Geometry.cs
@@ -2,7 +2,7 @@
 
 using OasysUnits;
 
-namespace AdSecCore.Helpers {
+namespace AdSecCore.Builders {
   public static class Geometry {
     public static IPoint Zero() {
       return IPoint.Create(Length.Zero, Length.Zero);
