+++ AdSecCore/DoubleComparer.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace AdSecGHCore.Helpers {
+namespace AdSecCore {
   public class DoubleComparer : IEqualityComparer<double> {
     private readonly double _epsilon;
     private readonly bool _margin;
