+++ AdSecCoreTests/FlattenRebarTests.cs
@@ -1,7 +1,6 @@
-﻿using AdSecCore.Builders;
-using AdSecCore.Helpers;
-
-using AdSecGH.Components;
+﻿using AdSecCore;
+using AdSecCore.Builders;
+using AdSecCore.Functions;
 
 using AdSecGHCore.Constants;
