+++ AdSecGHTests/Components/FlattenRebarGhTests.cs
@@ -1,5 +1,4 @@
 ﻿using AdSecCore.Builders;
-using AdSecCore.Helpers;
 
 using AdSecGH.Components;
 using AdSecGH.Parameters;
