+++ AdSecGHTests/Helpers/AdvancedDataConvertorTest.cs
@@ -1,11 +1,12 @@
 ﻿using System;
 
+using AdSecCore.Functions;
+
 using AdSecGH.Parameters;
 
 using Grasshopper.Kernel;
 using Grasshopper.Kernel.Parameters;
 
-using Oasys.Business;
 using Oasys.GH.Helpers;
 
 using Rhino.Geometry;
