+++ AdSecGHTests/Helpers/FlattenRebarGhTests.cs
@@ -1,4 +1,6 @@
-﻿using AdSecGH.Components;
+﻿using AdSecCore;
+
+using AdSecGH.Components;
 
 using OasysGH.Units;
